task 3

+++ 0x10-variadic_functions/3-print_all.c
@@ -3,14 +3,14 @@
 #include <stdio.h>
 
 /**
- * print_all - function that prints anything
- * @format: list of types of argumentspassed to the function
+ * print_all - prints anything
+ * @format: list of types of arguments passed to the function
  */
-
 void print_all(const char * const format, ...)
 {
 	int i = 0;
 	char *str, *sep = "";
+
 	va_list list;
 
 	va_start(list, format);
@@ -28,7 +28,7 @@ void print_all(const char * const format, ...)
 					printf("%s%d", sep, va_arg(list, int));
 					break;
 				case 'f':
-					printf("%s%d", sep, va_arg(list, int));
+					printf("%s%f", sep, va_arg(list, double));
 					break;
 				case 's':
 					str = va_arg(list, char *);
@@ -48,4 +48,3 @@ void print_all(const char * const format, ...)
 	printf("\n");
 	va_end(list);
 }
-
